+++ app/src/main/java/com/kiminonawa/mydiary/backup/ExportAsyncTask.java
@@ -16,7 +16,7 @@ import com.kiminonawa.mydiary.backup.obj.BUDiaryItem;
 import com.kiminonawa.mydiary.backup.obj.BUMemoEntries;
 import com.kiminonawa.mydiary.db.DBManager;
 import com.kiminonawa.mydiary.main.topic.ITopic;
-import com.kiminonawa.mydiary.shared.FileManager;
+import com.kiminonawa.mydiary.shared.file.FileManager;
 
 import java.io.FileWriter;
 import java.io.IOException;
