+++ app/src/main/java/com/kiminonawa/mydiary/main/YourNameDialogFragment.java
@@ -18,12 +18,13 @@ import android.widget.LinearLayout;
 import android.widget.Toast;
 
 import com.kiminonawa.mydiary.R;
-import com.kiminonawa.mydiary.shared.FileManager;
 import com.kiminonawa.mydiary.shared.PermissionHelper;
 import com.kiminonawa.mydiary.shared.SPFManager;
 import com.kiminonawa.mydiary.shared.ScreenHelper;
 import com.kiminonawa.mydiary.shared.ThemeManager;
 import com.kiminonawa.mydiary.shared.ViewTools;
+import com.kiminonawa.mydiary.shared.file.FileManager;
+import com.kiminonawa.mydiary.shared.file.MyDiaryFileUtils;
 import com.kiminonawa.mydiary.shared.gui.MyDiaryButton;
 import com.yalantis.ucrop.UCrop;
 
@@ -122,7 +123,7 @@ public class YourNameDialogFragment extends DialogFragment implements View.OnCli
         if (requestCode == PermissionHelper.REQUEST_WRITE_ES_PERMISSION) {
             if (grantResults.length > 0
                     && PermissionHelper.checkAllPermissionResult(grantResults)) {
-                FileManager.startBrowseImageFile(this, SELECT_PROFILE_PICTURE_BG);
+                MyDiaryFileUtils.startBrowseImageFile(this, SELECT_PROFILE_PICTURE_BG);
             } else {
                 PermissionHelper.showAddPhotoDialog(getActivity());
             }
@@ -144,7 +145,7 @@ public class YourNameDialogFragment extends DialogFragment implements View.OnCli
                     options.setToolbarColor(ThemeManager.getInstance().getThemeMainColor(getActivity()));
                     options.setStatusBarColor(ThemeManager.getInstance().getThemeDarkColor(getActivity()));
                     UCrop.of(data.getData(), Uri.fromFile(
-                            new File(tempFileManager.getDir() + "/" + FileManager.createRandomFileName())))
+                            new File(tempFileManager.getDir() + "/" + MyDiaryFileUtils.createRandomFileName())))
                             .withMaxResultSize(photoSize, photoSize)
                             .withAspectRatio(1, 1)
                             .withOptions(options)
@@ -158,7 +159,7 @@ public class YourNameDialogFragment extends DialogFragment implements View.OnCli
                 if (data != null) {
                     final Uri resultUri = UCrop.getOutput(data);
                     IV_your_name_profile_picture.setImageBitmap(BitmapFactory.decodeFile(resultUri.getPath()));
-                    profilePictureFileName = FileManager.getFileNameByUri(getActivity(), resultUri);
+                    profilePictureFileName = MyDiaryFileUtils.getFileNameByUri(getActivity(), resultUri);
                     isAddNewProfilePicture = true;
                 } else {
                     Toast.makeText(getActivity(), getString(R.string.toast_crop_profile_picture_fail), Toast.LENGTH_LONG).show();
@@ -189,7 +190,7 @@ public class YourNameDialogFragment extends DialogFragment implements View.OnCli
             if (!"".equals(profilePictureFileName)) {
                 try {
                     //Copy the profile into setting dir
-                    FileManager.copy(
+                    MyDiaryFileUtils.copy(
                             new File(tempFileManager.getDirAbsolutePath() + "/" + profilePictureFileName),
                             oldProfilePictureFile);
                 } catch (Exception e) {
@@ -206,7 +207,7 @@ public class YourNameDialogFragment extends DialogFragment implements View.OnCli
         switch (v.getId()) {
             case R.id.IV_your_name_profile_picture:
                 if (PermissionHelper.checkPermission(this, REQUEST_WRITE_ES_PERMISSION)) {
-                    FileManager.startBrowseImageFile(this, SELECT_PROFILE_PICTURE_BG);
+                    MyDiaryFileUtils.startBrowseImageFile(this, SELECT_PROFILE_PICTURE_BG);
                 }
                 break;
             case R.id.IV_your_name_profile_picture_cancel:
