+++ app/src/main/java/com/kiminonawa/mydiary/shared/file/FileFactory.java
@@ -1,0 +1,10 @@
+package com.kiminonawa.mydiary.shared.file;
+
+/**
+ * Created by daxia on 2018/2/8.
+ */
+
+public class FileFactory {
+
+
+}
