+++ app/src/main/java/com/kiminonawa/mydiary/shared/file/IFile.java
@@ -1,0 +1,8 @@
+package com.kiminonawa.mydiary.shared.file;
+
+/**
+ * Created by daxia on 2018/2/8.
+ */
+
+public class IFile {
+}
